requirements.txt
pandas
openpyxl 
camelot-py[cv] 